fix: move to base Serializable object
