+++ code_pylint.py
@@ -28,8 +28,8 @@ MAX_ERROR_BY_TYPE = {
                      'too-many-branches': 12,
                      'wrong-import-order': 1,
                      'too-many-branches': 12,
-                     'unused-argument': 5,
-                     'cyclic-import': 12,
+                     'unused-argument': 6,
+                     'cyclic-import': 14,
                      'no-self-use': 6,
                      'trailing-whitespace': 11,
                      'empty-docstring': 1,
