+++ dessia_common/core.py
@@ -6,7 +6,6 @@ dessia_common
 """
 
 import time
-import sys
 import warnings
 import operator
 import math
@@ -20,7 +19,7 @@ from copy import deepcopy, copy
 import inspect
 import json
 
-from typing import List, Dict, Any, Tuple, get_type_hints
+from typing import List, Tuple, get_type_hints
 import traceback as tb
 
 from importlib import import_module
@@ -28,17 +27,18 @@ from ast import literal_eval
 
 import dessia_common.errors
 from dessia_common.utils.diff import data_eq, diff, dict_hash, list_hash
-from dessia_common.utils.serialization import dict_to_object, serialize_dict_with_pointers, serialize_dict,\
-    deserialize_argument, serialize
+from dessia_common.utils.serialization import deserialize_argument, serialize
 from dessia_common.utils.types import full_classname, is_sequence, is_bson_valid, TYPES_FROM_STRING
 from dessia_common.utils.copy import deepcopy_value
 from dessia_common.utils.jsonschema import default_dict, jsonschema_from_annotation, JSONSCHEMA_HEADER,\
     set_default_value
 from dessia_common.utils.docstrings import parse_docstring, FAILED_DOCSTRING_PARSING
+
+from dessia_common.base import SerializableObject
 from dessia_common.exports import XLSXWriter
 from dessia_common.typings import JsonSerializable
 from dessia_common import templates
-from dessia_common.checks import type_check, CheckList
+from dessia_common.checks import CheckList, FailedCheck
 from dessia_common.displays import DisplayObject, DisplaySetting
 from dessia_common.breakdown import attrmethod_getter, get_in_object_from_path
 
@@ -66,78 +66,6 @@ def deprecation_warning(name, object_type, use_instead=None):
         msg += f"Use {use_instead} instead.\n"
     warnings.warn(msg, DeprecationWarning)
     return msg
-
-
-class SerializableObject:
-    """
-    Serialization capabilities of Dessia Object
-    """
-    _non_serializable_attributes = []
-
-    def base_dict(self):
-        """
-        A base dict for to_dict: put name, object class and version in a dict
-        """
-        package_name = self.__module__.split('.', maxsplit=1)[0]
-        if package_name in sys.modules:
-            package = sys.modules[package_name]
-            if hasattr(package, '__version__'):
-                package_version = package.__version__
-            else:
-                package_version = None
-        else:
-            package_version = None
-
-        object_class = self.__module__ + '.' + self.__class__.__name__
-        dict_ = {'object_class': object_class}
-        if package_version:
-            dict_['package_version'] = package_version
-        return dict_
-
-    def _serializable_dict(self):
-        """
-        Returns a dict of attribute_name, values (still python, not serialized)
-        Keys are filtered with non serializable attributes controls
-        """
-
-        dict_ = {k: v for k, v in self.__dict__.items()
-                 if k not in self._non_serializable_attributes and not k.startswith('_')}
-        return dict_
-
-    def to_dict(self, use_pointers: bool = True, memo=None, path: str = '#') -> JsonSerializable:
-        """
-        Generic to_dict method
-        """
-        if memo is None:
-            memo = {}
-
-        # Default to dict
-        serialized_dict = self.base_dict()
-        dict_ = self._serializable_dict()
-        if use_pointers:
-            serialized_dict.update(serialize_dict_with_pointers(dict_, memo, path)[0])
-        else:
-            serialized_dict.update(serialize_dict(dict_))
-
-        return serialized_dict
-
-    @classmethod
-    def dict_to_object(cls, dict_: JsonSerializable, force_generic: bool = False, global_dict=None,
-                       pointers_memo: Dict[str, Any] = None, path: str = '#') -> 'DessiaObject':
-        """
-        Generic dict_to_object method
-        """
-        if cls is not DessiaObject:
-            obj = dict_to_object(dict_=dict_, class_=cls, force_generic=force_generic, global_dict=global_dict,
-                                 pointers_memo=pointers_memo, path=path)
-            return obj
-
-        if 'object_class' in dict_:
-            obj = dict_to_object(dict_=dict_, force_generic=force_generic, global_dict=global_dict,
-                                 pointers_memo=pointers_memo, path=path)
-            return obj
-
-        raise NotImplementedError('No object_class in dict')
 
 
 class DessiaObject(SerializableObject):
@@ -457,17 +385,22 @@ class DessiaObject(SerializableObject):
         return cls.dict_to_object(dict_)
 
     def checks(self, level='error'):
-        class_argspec = inspect.getfullargspec(self.__class__)
-        annotations = inspect.signature(self.__init__).parameters
         check_list = CheckList([])
 
-        for arg in class_argspec.args:
-            if arg != 'self':
-                if arg in annotations:
-                    value = self.__dict__[arg]
-                    # print(annotations[arg], type(annotations[arg]))
-                    print(annotations[arg].annotation)
-                    check_list += type_check(value, annotations[arg].annotation.__class__, level=level)
+        check_list += self._check_platform(level=level)
+
+        # Type checking: not ready yet
+        # class_argspec = inspect.getfullargspec(self.__class__)
+        # annotations = inspect.signature(self.__init__).parameters
+
+        # for arg in class_argspec.args:
+        #     if arg != 'self':
+        #         if arg in annotations:
+        #             value = self.__dict__[arg]
+        #             # print(annotations[arg], type(annotations[arg]))
+        #             print(annotations[arg].annotation)
+        #             check_list += type_check(value, annotations[arg].annotation.__class__, level=level)
+
         return check_list
 
     def copy(self, deep=True, memo=None):
@@ -616,11 +549,12 @@ class DessiaObject(SerializableObject):
             display_time = time.time() - display_time
             print(f'Generation of display {display_setting.selector} in: {round(display_time, 6)} seconds')
 
-    def _check_platform(self):
+    def _check_platform(self, level='error'):
         """
         Reproduce lifecycle on platform (serialization, display)
         raise an error if something is wrong
         """
+        checks = []
         try:
             dict_ = self.to_dict(use_pointers=True)
         except TypeError:
@@ -630,21 +564,27 @@ class DessiaObject(SerializableObject):
         deserialized_object = self.dict_to_object(decoded_json)
         if not deserialized_object._data_eq(self):
             print('data diff: ', self._data_diff(deserialized_object))
-            raise dessia_common.errors.DeserializationError('Object is not equal to itself'
-                                                            ' after serialization/deserialization')
+            # raise dessia_common.errors.DeserializationError('Object is not equal to itself'
+            #                                                 ' after serialization/deserialization')
+            checks.append(FailedCheck('Object is not equal to itself after serialization/deserialization'))
         copied_object = self.copy()
         if not copied_object._data_eq(self):
             try:
                 print('data diff: ', self._data_diff(copied_object))
             except:
                 pass
-            raise dessia_common.errors.CopyError('Object is not equal to itself after copy')
+            checks.append(FailedCheck('Object is not equal to itself after serialization/deserialization'))
+            # raise dessia_common.errors.CopyError('Object is not equal to itself after copy')
 
         valid, hint = is_bson_valid(stringify_dict_keys(dict_))
         if not valid:
-            raise ValueError(hint)
+            # raise ValueError(hint)
+            checks.append(FailedCheck(f'Object is not bson valid {hint}'))
+
         json.dumps(self._displays())
         json.dumps(self._method_jsonschemas)
+
+        return CheckList(checks)
 
     def to_xlsx(self, filepath: str):
         """
